requirements.txt
groq
watchdog
boto3
python-dotenv